\name{GANPA-package}
\alias{GANPA-package}
\alias{GANPA}
\docType{package}
\title{Gene Association Network-based Pathway Analysis}
\description{
This package implements a network-based gene weighting algorithm for pathways, as well as a gene-weighted gene set analysis approach for microarray data pathway analysis.
}
\details{
\tabular{ll}{
Package: \tab GANPA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-05-24\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Zhaoyuan Fang, Weidong Tian and Hongbin Ji

Maintainer: Zhaoyuan Fang \email{fangzhaoyuan@sibs.ac.cn}
}
\references{
Zhaoyuan Fang, Weidong Tian and Hongbin Ji. A Network-Based Gene Weighting Approach for Pathway Analysis. \emph{Submitted}.

Aravind Subramanian, Pablo Tamayo, Vamsi K. Mootha, Sayan Mukherjee, Benjamin L. Ebert, Michael A. Gillette, Amanda Paulovich, Scott L. Pomeroy, Todd R. Golub, Eric S. Lander, and Jill P. Mesirov. Gene set enrichment analysis: A knowledge-based approach for interpreting genome-wide expression profiles. \emph{PNAS} 2005 102(43): 15545-15550. 
}
\keyword{package}
\seealso{}
\examples{}
